requirements.txt
aiohttp==3.7.4.post0
amqp==5.0.6
asgiref==3.3.4
async-timeout==3.0.1
attrs==20.3.0
billiard==3.6.4.0
celery==5.1.0b1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cryptography==3.4.7
curlify==2.2.1
Django==3.2.3
django-appconf==1.0.4
django-background-tasks==1.2.5
django-bulk-sync==3.2.0
django-celery-beat==2.2.0
django-celery-results==2.0.1
django-compat==1.0.15
django-cors-headers==3.7.0
django-cryptography==1.0
django-debug-toolbar==3.2.1
django-timezone-field==4.1.2
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
drf-nested-routers==0.93.3
facebook-business==10.0.0
facebook-sdk==3.1.0
gevent==21.1.2
greenlet==1.0.0
idna==2.10
kombu==5.1.0b1
multidict==5.1.0
numpy==1.20.2
pandas==1.2.4
prompt-toolkit==3.0.18
psycopg2==2.8.6
pycountry==20.7.3
pycparser==2.20
PyJWT==2.0.1
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
urllib3==1.26.4
vine==5.0.0
wcwidth==0.2.5
yarl==1.6.3
zope.event==4.5.0
zope.interface==5.4.0